fixed formatting for listing pages
