+++ src/components/ListingContractQuestions/ListingCategory6.js
@@ -5,48 +5,49 @@ import { withRouter } from 'react-router';
 import Grid from '@material-ui/core/Grid';
 import Checkbox from '@material-ui/core/Checkbox';
 import Container from '@material-ui/core/Container';
+import { TextField } from '@material-ui/core';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Button from '@material-ui/core/Button';
 
 
-
-class ListingCategory7 extends Component {
-
+class ListingCategory6 extends Component {
     state = {
-        foreignInvestment: {
-            sellerRepYes:'',
-            sellerRepNo:'',
+        closingServices: {
+            arrangeClosing: '',
+            conductClosing: '',
+            initials: '',
         }
     }
 
+
     handleClickForCheckBox = (propertyName) => (event) => {
         this.setState({
-            foreignInvestment: {
-                ...this.state.foreignInvestment, [propertyName]:true
+            closingServices: {
+                ...this.state.closingServices, [propertyName]: true,
             }
         })
     }
 
-    handleClickForCheckBoxNo = (propertyName) => (event) => {
+    handleChangeForInitials = (propertyName) => (event) => {
         this.setState({
-            foreignInvestment: {
-                ...this.state.foreignInvestment, [propertyName]:false
+            closingServices: {
+                ...this.state.closingServices, [propertyName]: event.target.value,
             }
         })
     }
-   
+
 
     handleClick = () => {
         this.props.history.push('/ListingContract')
     }
 
     handleClickToSave = () => {
-        this.props.dispatch({type:'POST_FOREIGN_INVESTMENT', payload:this.state.foreignInvestment});
-        this.props.history.push('/ListingContract')
+        this.props.dispatch({ type: 'CLOSING_SERVICES', payload: this.state.closingServices });
+        this.props.history.push('/ListingContract');
     }
 
     handleClickNext = () => {
-        this.props.history.push('/ListingCategory8')
+        this.props.history.push('/ListingCategory7')
     }
 
 
@@ -55,12 +56,22 @@ class ListingCategory7 extends Component {
             <div>
                 <Container component="main">
                     <Grid container spacing={2}>
+
                         <Grid item xs={12}>
                             <center>
-                                <h2>Foreign Investment</h2>
+                                <h2>Closing Services</h2>
                             </center>
+                            <h4>Seller's choice for closing services (Check one):</h4>
+                            <FormControlLabel
+                                control={
+                                    <Checkbox
+                                        value="checkedB"
+                                        color="primary" />
+                                }
+                                label="Broker will arrange for the closing."
+                                onClick={this.handleClickForCheckBox('arrangeClosing')}
+                            />
                         </Grid>
-                        <h4>Seller represents and warrants that seller is a foreign person for purpose of income taxation</h4>
                         <Grid item xs={12}>
                             <FormControlLabel
                                 control={
@@ -68,26 +79,27 @@ class ListingCategory7 extends Component {
                                         value="checkedB"
                                         color="primary" />
                                 }
-                                label="Yes"
-                                onClick={this.handleClickForCheckBox('sellerRepYes')}
-                            />
-                            <FormControlLabel
-                                control={
-                                    <Checkbox
-                                        value="checkedB"
-                                        color="primary" />
-                                }
-                                label="No"
-                                onClick={this.handleClickForCheckBoxNo('sellerRepNo')}
+                                label="Qualified agent or Seller's attorney will conduct the closing."
+                                onClick={this.handleClickForCheckBox('conductClosing')}
                             />
                         </Grid>
-                        <p>Due to the complexity and potential risks of failing to comply with FIRPTA, Seller should <b>seek appropiate legal and tax advice regarding FIRPTA compliance, as Broker will be unable to confirm whether Seeker is a foreign person or whether the withholding requirements of FIRPTA apply.</b> </p>
+                        <br />
+                        <br />
+                        <Grid item xs={12}>
+                            <TextField
+                                label="Seller's Initials"
+                                fullWidth
+                                variant="outlined"
+                                value={this.setState.initials}
+                                onChange={this.handleChangeForInitials('initials')}
+                            />
+                        </Grid>
                     </Grid>
                 </Container>
                 <br />
                 <br />
                 <Container component="main">
-                    <Grid container spacing={2}>
+                    <Grid container spacing={3}>
                         <Grid item xs={4}>
                             <div align="left" className="Button">
                                 <Button
@@ -126,10 +138,10 @@ class ListingCategory7 extends Component {
                 <pre>
                     {JSON.stringify(this.state, null, 2)}
                 </pre>
-            </div >
+            </div>
         )
     }
 }
 
 const mapReduxStateToProps = reduxState => ({ reduxState })
-export default connect(mapReduxStateToProps)(withRouter(ListingCategory7));
+export default connect(mapReduxStateToProps)(withRouter(ListingCategory6));
